fix(api): Introduce Universal Job Status Monitoring with Self-Cleaning Subscriptions
