+++ server/api/internal/usecase/interactor/job.go
@@ -31,8 +31,8 @@ type Job struct {
 	subscriptions     *subscription.JobManager
 	notifier          notification.Notifier
 	permissionChecker gateway.PermissionChecker
-	watchersMu        sync.Mutex
-	activeWatchers    map[string]bool
+	monitoringMu      sync.RWMutex
+	monitoringJobs    map[string]time.Time
 }
 
 type NotificationPayload struct {
@@ -54,8 +54,8 @@ func NewJob(r *repo.Container, gr *gateway.Container, permissionChecker gateway.
 		subscriptions:     subscription.NewJobManager(),
 		notifier:          notification.NewHTTPNotifier(),
 		permissionChecker: permissionChecker,
-		watchersMu:        sync.Mutex{},
-		activeWatchers:    make(map[string]bool),
+		monitoringMu:      sync.RWMutex{},
+		monitoringJobs:    make(map[string]time.Time),
 	}
 }
 
@@ -153,290 +153,4 @@ func (i *Job) GetStatus(ctx context.Context, jobID id.JobID) (job.Status, error)
 		return "", err
 	}
 	return j.Status(), nil
-}
-
-func (i *Job) StartMonitoring(ctx context.Context, j *job.Job, notificationURL *string) error {
-	if err := i.checkPermission(ctx, rbac.ActionAny); err != nil {
-		return err
-	}
-
-	log.Infof("starting monitoring for job ID %s with workspace %s", j.ID(), j.Workspace())
-
-	monitorCtx, cancel := context.WithCancel(context.Background())
-
-	i.monitor.Register(j.ID().String(), &monitor.Config{
-		Cancel:          cancel,
-		NotificationURL: notificationURL,
-	})
-
-	go i.runMonitoringLoop(monitorCtx, j)
-
-	return nil
-}
-
-func (i *Job) runMonitoringLoop(ctx context.Context, j *job.Job) {
-	ticker := time.NewTicker(10 * time.Second)
-	defer ticker.Stop()
-
-	jobID := j.ID().String()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-ticker.C:
-			i.watchersMu.Lock()
-			if i.subscriptions.CountSubscribers(jobID) == 0 {
-				delete(i.activeWatchers, jobID)
-				i.watchersMu.Unlock()
-
-				i.monitor.Remove(jobID)
-				return
-			}
-			i.watchersMu.Unlock()
-
-			if err := i.checkJobStatus(ctx, j); err != nil {
-				log.Errorfc(ctx, "job: status check failed: %v", err)
-			}
-		}
-	}
-}
-
-func (i *Job) checkJobStatus(ctx context.Context, j *job.Job) error {
-	if err := i.checkPermission(ctx, rbac.ActionAny); err != nil {
-		return err
-	}
-
-	status, err := i.batch.GetJobStatus(ctx, j.GCPJobID())
-	if err != nil {
-		return err
-	}
-
-	newStatus := job.Status(status)
-	statusChanged := j.Status() != newStatus
-
-	if status == gateway.JobStatusCompleted || status == gateway.JobStatusFailed {
-		if statusChanged {
-			log.Infof("job status changing to terminal state %s from %s for job ID %s",
-				newStatus, j.Status(), j.ID())
-
-			if err := i.updateJobStatus(ctx, j, newStatus); err != nil {
-				return err
-			}
-
-			if err := i.handleJobCompletion(ctx, j); err != nil {
-				log.Errorfc(ctx, "job: completion handling failed: %v", err)
-			}
-
-			i.monitor.Remove(j.ID().String())
-		} else {
-			if len(j.OutputURLs()) == 0 {
-				log.Infof("job already in terminal state %s but has no outputs, checking artifacts for job ID %s",
-					j.Status(), j.ID())
-
-				if err := i.handleJobCompletion(ctx, j); err != nil {
-					log.Errorfc(ctx, "job: completion handling failed: %v", err)
-				}
-			}
-		}
-	} else if statusChanged {
-		log.Infof("job status changing from %s to %s for job ID %s",
-			j.Status(), newStatus, j.ID())
-
-		if err := i.updateJobStatus(ctx, j, newStatus); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (i *Job) updateJobStatus(ctx context.Context, j *job.Job, status job.Status) error {
-	if err := i.checkPermission(ctx, rbac.ActionAny); err != nil {
-		return err
-	}
-
-	tx, err := i.transaction.Begin(ctx)
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if err := tx.End(ctx); err != nil {
-			log.Errorfc(ctx, "transaction end failed: %v", err)
-		}
-	}()
-
-	j.SetStatus(status)
-	if err := i.jobRepo.Save(ctx, j); err != nil {
-		return err
-	}
-
-	tx.Commit()
-	i.subscriptions.Notify(j.ID().String(), j.Status())
-	return nil
-}
-
-func (i *Job) handleJobCompletion(ctx context.Context, j *job.Job) error {
-	if j == nil {
-		return fmt.Errorf("job cannot be nil")
-	}
-
-	if err := i.checkPermission(ctx, rbac.ActionAny); err != nil {
-		return err
-	}
-
-	jobID := j.ID().String()
-	log.Infof("handling job completion for ID %s with status %s", jobID, j.Status())
-
-	config := i.monitor.Get(jobID)
-
-	var artifactErr error
-	for retries := 0; retries < 3; retries++ {
-		if retries > 0 {
-			log.Infof("retry %d: updating artifacts for job ID %s", retries, jobID)
-			time.Sleep(2 * time.Second)
-		}
-
-		artifactErr = i.updateJobArtifacts(ctx, j)
-		if artifactErr == nil && len(j.OutputURLs()) > 0 {
-			break
-		}
-	}
-
-	if artifactErr != nil {
-		log.Errorfc(ctx, "job: failed to update artifacts after retries for jobID=%s: %v", jobID, artifactErr)
-	}
-
-	if err := i.saveJobState(ctx, j); err != nil {
-		return fmt.Errorf("failed to save job state: %w", err)
-	}
-
-	if config == nil || config.NotificationURL == nil || *config.NotificationURL == "" {
-		return nil
-	}
-
-	if err := i.sendCompletionNotification(ctx, j, *config.NotificationURL); err != nil {
-		return fmt.Errorf("failed to send notification: %w", err)
-	}
-
-	return nil
-}
-
-func (i *Job) updateJobArtifacts(ctx context.Context, j *job.Job) error {
-	jobID := j.ID().String()
-
-	outputs, err := i.file.ListJobArtifacts(ctx, jobID)
-	if err != nil {
-		return fmt.Errorf("failed to list job artifacts: %w", err)
-	}
-	j.SetOutputURLs(outputs)
-
-	logURL := i.file.GetJobLogURL(jobID)
-	if logURL != "" {
-		j.SetLogsURL(logURL)
-	}
-
-	return nil
-}
-
-func (i *Job) saveJobState(ctx context.Context, j *job.Job) error {
-	tx, err := i.transaction.Begin(ctx)
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction: %w", err)
-	}
-
-	defer func() {
-		if err := tx.End(ctx); err != nil {
-			log.Errorfc(ctx, "transaction end failed: %v", err)
-		}
-	}()
-
-	if err := i.jobRepo.Save(ctx, j); err != nil {
-		return fmt.Errorf("failed to save job: %w", err)
-	}
-
-	tx.Commit()
-	return nil
-}
-
-func (i *Job) sendCompletionNotification(ctx context.Context, j *job.Job, notificationURL string) error {
-	jobID := j.ID().String()
-
-	status := "failed"
-	switch j.Status() {
-	case job.StatusCompleted:
-		status = "succeeded"
-	case job.StatusCancelled:
-		status = "cancelled"
-	}
-
-	var logs []string
-	logExists, err := i.file.CheckJobLogExists(ctx, jobID)
-	if err != nil {
-		log.Warnfc(ctx, "job: failed to check log existence for jobID=%s: %v", jobID, err)
-	} else if logExists {
-		logs = append(logs, j.LogsURL())
-	}
-
-	payload := notification.Payload{
-		RunID:        jobID,
-		DeploymentID: j.Deployment().String(),
-		Status:       status,
-		Logs:         logs,
-		Outputs:      j.OutputURLs(),
-	}
-
-	return i.notifier.Send(notificationURL, payload)
-}
-
-func (i *Job) Subscribe(ctx context.Context, jobID id.JobID) (chan job.Status, error) {
-	if err := i.checkPermission(ctx, rbac.ActionAny); err != nil {
-		return nil, err
-	}
-
-	ch := i.subscriptions.Subscribe(jobID.String())
-
-	go func() {
-		j, err := i.FindByID(context.Background(), jobID)
-		if err == nil {
-			i.subscriptions.Notify(jobID.String(), j.Status())
-		}
-	}()
-
-	i.startMonitoringIfNeeded(jobID)
-
-	return ch, nil
-}
-
-func (i *Job) startMonitoringIfNeeded(jobID id.JobID) {
-	i.watchersMu.Lock()
-	defer i.watchersMu.Unlock()
-
-	jobKey := jobID.String()
-	if i.activeWatchers == nil {
-		i.activeWatchers = make(map[string]bool)
-	}
-
-	if _, exists := i.activeWatchers[jobKey]; exists {
-		return
-	}
-
-	i.activeWatchers[jobKey] = true
-
-	j, err := i.jobRepo.FindByID(context.Background(), jobID)
-	if err != nil {
-		log.Errorfc(context.Background(), "job: failed to find job for monitoring: %v", err)
-		return
-	}
-
-	monitorCtx, cancel := context.WithCancel(context.Background())
-	i.monitor.Register(jobKey, &monitor.Config{
-		Cancel: cancel,
-	})
-
-	go i.runMonitoringLoop(monitorCtx, j)
-}
-
-func (i *Job) Unsubscribe(jobID id.JobID, ch chan job.Status) {
-	i.subscriptions.Unsubscribe(jobID.String(), ch)
 }
