+++ server/api/internal/usecase/interactor/trigger.go
@@ -164,6 +164,9 @@ func (i *Trigger) ExecuteAPITrigger(ctx context.Context, p interfaces.ExecuteAPI
 	}
 
 	metadataURL, err := i.file.UploadMetadata(ctx, j.ID().String(), []string{})
+	if err != nil {
+		return nil, err
+	}
 	j.SetMetadataURL(metadataURL.String())
 	if err := i.jobRepo.Save(ctx, j); err != nil {
 		return nil, err
@@ -174,29 +177,24 @@ func (i *Trigger) ExecuteAPITrigger(ctx context.Context, p interfaces.ExecuteAPI
 		projectID = *deployment.Project()
 	}
 
-	log.Debugfc(ctx, "[Trigger] Submitting job ID: %s, workflow: %s\n", j.ID(), deployment.WorkflowURL())
-
 	gcpJobID, err := i.batch.SubmitJob(ctx, j.ID(), deployment.WorkflowURL(), j.MetadataURL(), p.Variables, projectID, deployment.Workspace())
 	if err != nil {
 		log.Debugfc(ctx, "[Trigger] Job submission failed: %v\n", err)
 		return nil, interfaces.ErrJobCreationFailed
 	}
-	log.Debugfc(ctx, "[Trigger] GCP job ID received: %s\n", gcpJobID)
-	if gcpJobID == "" {
-		log.Debugfc(ctx, "[Trigger] WARNING: Empty GCP job ID returned\n")
-	} else {
-		j.SetGCPJobID(gcpJobID)
-		if err := i.jobRepo.Save(ctx, j); err != nil {
-			log.Debugfc(ctx, "[Trigger] Failed to save job with GCP ID: %v\n", err)
-		}
+
+	j.SetGCPJobID(gcpJobID)
+	if err := i.jobRepo.Save(ctx, j); err != nil {
+		log.Errorf("Failed to save job %s with GCP ID: %v", j.ID(), err)
+		return nil, err
 	}
 
 	if err := i.job.StartMonitoring(ctx, j, p.NotificationURL); err != nil {
+		log.Errorf("Failed to start monitoring for job %s: %v", j.ID(), err)
 		return nil, err
 	}
 
 	tx.Commit()
-
 	return j, nil
 }
